+++ src/anime_bridge/__init__.py
@@ -1,3 +1,3 @@
 """Anime Bridge public package metadata."""
 
-__version__ = "0.5.0.dev0"
+__version__ = "0.6.0.dev0"